# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CUYHUB API",
        "version": "0.1"
    },
    "paths": {
        "/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login User",
                "description": "login user.",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "schema": "LoginRequest",
                                "title": "LoginRequest",
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "get json result with token"
                    },
                    "422": {
                        "description": "there is a body that is needed there is not"
                    },
                    "401": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/v1/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register User",
                "description": "register user.",
                "operationId": "createUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "schema": "RegisterRequest",
                                "title": "RegisterRequest",
                                "required": [
                                    "fullName",
                                    "username",
                                    "password",
                                    "password_confirmation",
                                    "email"
                                ],
                                "properties": {
                                    "fullName": {
                                        "type": "string"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "get result form-data users"
                    },
                    "422": {
                        "description": "there is a body that is needed there is not"
                    },
                    "401": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/v1/profile/": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get Profile",
                "operationId": "a7ef021f26fa3d25baaeb3b7b8dd5f9b",
                "responses": {
                    "200": {
                        "description": "get profile"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/users/?_method=PATCH": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update profile",
                "operationId": "dc7779e3c3d55869ca44194ae26aa3be",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "schema": "UpdateProfileRequest",
                                "title": "UpdateProfileRequest",
                                "required": [
                                    "fullName",
                                    "username",
                                    "email"
                                ],
                                "properties": {
                                    "fullName": {
                                        "type": "string"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "get result form-data users"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "400": {
                        "description": "Failed update data user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/users/": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List users",
                "operationId": "4bd22def501d28e780b2fb520e7e97e4",
                "responses": {
                    "200": {
                        "description": "A list with users"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Add User",
                "description": "add user.",
                "operationId": "9eece8801ce21f5442413ac712b6fe0f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "schema": "AddUserRequest",
                                "title": "AddUserRequest",
                                "required": [
                                    "fullName",
                                    "username",
                                    "password",
                                    "password_confirmation",
                                    "email"
                                ],
                                "properties": {
                                    "fullName": {
                                        "type": "string"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Get result form-data users"
                    },
                    "422": {
                        "description": "There is a body that is needed there is not"
                    },
                    "401": {
                        "description": "Validation Error or Unauthorized"
                    },
                    "400": {
                        "description": "Failed insert data user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user by ID",
                "operationId": "a59a43efe8853a85f0adcb09dd6163c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID user",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get information user by id user"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete user by ID",
                "operationId": "4d460066ae6ef41c5345f855c3bc708d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID user",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete user by ID"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/users/{id}?_method=PATCH": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user by ID",
                "operationId": "d757c0ad85a75c8fa11b2d3327d701a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID user",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "schema": "UpdateUserRequest",
                                "title": "UpdateUserRequest",
                                "required": [
                                    "fullName",
                                    "username",
                                    "email"
                                ],
                                "properties": {
                                    "fullName": {
                                        "type": "string"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "get result form-data users"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "400": {
                        "description": "Failed update data user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}